+++ src/commands.h
@@ -6,13 +6,13 @@
 
 class Commands {
 public:
-    std::string login(uint8_t tag, std::string& username, const std::string& password);
-    std::string logout(uint8_t tag);
-    std::string list(uint8_t tag, const std::string& reference, const std::string& mailbox);
-    std::string select(uint8_t tag, const std::string& mailbox);
-    std::string fetch_header_important(uint8_t tag, const int message_id);
-    std::string fetch_header(uint8_t tag,  const int message_id);
-    std::string fetch(uint8_t tag, int message_id);
-    std::string get_new_messages(uint8_t tag);
+    std::string login(uint64_t tag, std::string& username, const std::string& password);
+    std::string logout(uint64_t tag);
+    std::string list(uint64_t tag, const std::string& reference, const std::string& mailbox);
+    std::string select(uint64_t tag, const std::string& mailbox);
+    std::string fetch_header_important(uint64_t tag, const int message_id);
+    std::string fetch_header(uint64_t tag,  const int message_id);
+    std::string fetch(uint64_t tag, int message_id);
+    std::string get_new_messages(uint64_t tag);
     
 };
